Add payment details screen and enhance UI elements
Introduced PaymentDetailsScreen and integrated it into the app drawer. Updated ContactUsModel to include websiteLink and added a website button to the Contact Us screen. Improved text styling and layout in MyRequestScreen, ContactUsScreen, and AppDrawer. Enhanced logout flow in HomeServiceView with a confirmation dialog. Added a registration instruction and adjusted spacing in VendorRegScreen.